Move source files under a 'home' directory
Move all source files destined for the target $HOME under
a new 'home' directory in the source tree.

Reference this new directory via a top level .chezmoiroot file.

This is to make it more obvious which files are source only,
such as README and directives to chezmoi itself, versus files
destined for the target.