Add mermaid CLI layer in mh-install
Note: there is very likely a simpler way to get this, or at
least more generalised, using some combination of 'npm install -g'
or npx.
The main aim of this approach is to avoid installing packages in
the global node_modules of a mise node install, as
(1) I want to keep the language installs emphemeral, and
(2) Updates to node would likely mean scattered/lost state
from the previous install.